temp

+++ src/command/kill.ts
@@ -1,15 +1,17 @@
-import { Command } from "@/core/scope";
-import { Coordinate } from "@/type/coord";
-import { TARGET } from "@/type/selector";
+import { Command } from "../core/scope";
+import { Coordinate } from "../type/coord";
+import { TARGET } from "../type/selector";
 
 
 
 class Kill extends Command {
-    constructor(private t: TARGET) {
+    constructor(private t?: TARGET) {
         super()
     }
     public toString(): string {
-        return `kill ${this.t}`
+        if(this.t)
+            return `kill ${this.t}`
+        return `kill`
     }
 }
